3.1.3.2 создаём const POSTERS_IMG и рандомно выводим poster у фильма
