+++ src/const.js
@@ -1,3 +1,13 @@
+export const POSTERS_IMG = [
+  `./images/posters/made-for-each-other.png`,
+  `./images/posters/popeye-meets-sinbad.png`,
+  `./images/posters/sagebrush-trail.jpg`,
+  `./images/posters/santa-claus-conquers-the-martians.jpg`,
+  `./images/posters/the-dance-of-life.jpg`,
+  `./images/posters/the-great-flamarion.jpg`,
+  `./images/posters/the-man-with-the-golden-arm.jpg`,
+];
+
 export const NAVIGATION_NAMES = [
   `All movies`, `Watchlist`, `History`, `Favorites`, `Stats`
 ];
@@ -5,3 +15,5 @@ export const NAVIGATION_NAMES = [
 export const SORT_NAMES = [
   `default`, `date`, `rating`
 ];
+
+export const DAYS = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
